# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Burnout Monitor API
  description: API for employee burnout detection and peak performance identification
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        department:
          type: string
        jobTitle:
          type: string
        zone:
          type: string
          enum: [red, yellow, green]
        burnoutScore:
          type: number
        readinessScore:
          type: number
...
    WorkMetrics:
      type: object
      properties:
        date:
          type: string
          format: date
        hoursWorked:
          type: number
        tasksCompleted:
          type: integer
        tasksAssigned:
          type: integer
        meetingsAttended:
          type: integer
        avgResponseTimeMinutes:
          type: integer
    ZoneStatus:
      type: object
      properties:
        zone:
          type: string
          enum: [red, yellow, green]
        burnoutScore:
          type: number
        readinessScore:
          type: number
        date:
          type: string
          format: date
    Explanation:
      type: object
      properties:
        zone:
          type: string
          enum: [red, yellow, green]
        factors:
          type: array
          items:
            $ref: '#/components/schemas/Factor'
        recommendations:
          type: array
          items:
            type: string
    Factor:
      type: object
      properties:
        name:
          type: string
        impact:
          type: string
          enum: [positive, negative, neutral]
        value:
          type: string
        description:
          type: string
        weight:
          type: number
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      tags: [Employees]
      summary: List all employees (manager only)
      description: Returns list of all employees with current zone status
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '403':
          description: Forbidden - manager role required
  /employees/{id}:
    get:
      tags: [Employees]
      summary: Get employee details
      description: Manager can view any employee; employee can only view self
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '403':
          description: Forbidden
        '404':
          description: Employee not found
...
  /employees/{id}/work:
    get:
      tags: [Employees]
      summary: Get employee work metrics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Work metrics array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkMetrics'
  /employees/{id}/burnout:
    get:
      tags: [Employees]
      summary: Get employee burnout score history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Burnout score history
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    $ref: '#/components/schemas/ZoneStatus'
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ZoneStatus'
  /employees/{id}/readiness:
    get:
      tags: [Employees]
      summary: Get employee readiness score history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Readiness score history
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    $ref: '#/components/schemas/ZoneStatus'
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ZoneStatus'
  /employees/{id}/explanation:
    get:
      tags: [Employees]
      summary: Get explanation for current zone status
      description: Returns factors contributing to the employee's current zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Zone explanation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Explanation'

  # ==========================================
  # Alerts
  # ==========================================
...